requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.15.0=pypi_0
adversarial-robustness-toolbox=1.14.0=pypi_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
ca-certificates=2022.10.11=haa95532_0
cachetools=5.2.1=pypi_0
certifi=2022.12.7=py38haa95532_0
charset-normalizer=3.0.1=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.5=py38h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
deprecated=1.2.14=pypi_0
fftw=3.3.9=h2bbff1b_1
flatbuffers=1.12=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
gast=0.3.3=pypi_0
glib=2.69.1=h5dc1a3c_2
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.32.0=pypi_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
h5py=2.10.0=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
intel-openmp=2021.4.0=haa95532_3556
joblib=1.1.1=py38haa95532_0
jpeg=9e=h2bbff1b_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=py38hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libtiff=4.5.0=h8a3f274_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.6.2=py38haa95532_0
matplotlib-base=3.6.2=py38h1094b8e_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
munkres=1.1.4=py_0
numpy=1.19.5=pypi_0
numpy-base=1.23.5=py38h4da318b_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=1.1.1s=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=22.0=py38haa95532_0
pandas=2.0.3=pypi_0
pcre=8.45=hd77b12b_0
pillow=9.3.0=py38hdc2b20a_1
pip=22.3.1=py38haa95532_0
ply=3.11=py38_0
protobuf=3.20.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.7=py38hd77b12b_0
pyqt5-sip=12.11.0=py38hd77b12b_0
python=3.8.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2023.3.post1=pypi_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.9.3=py38he11b74f_0
setuptools=65.6.3=py38haa95532_0
sip=6.6.2=py38hd77b12b_0
six=1.15.0=pypi_0
sqlite=3.40.1=h2bbff1b_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-estimator=2.4.0=pypi_0
tensorflow-gpu=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.2=py38h2bbff1b_0
tqdm=4.65.0=pypi_0
typing-extensions=3.7.4.3=pypi_0
tzdata=2023.4=pypi_0
urllib3=1.26.14=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
wrapt=1.12.1=pypi_0
xz=5.2.10=h8cc25b3_1
zipp=3.11.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
